ref(useDropGenerativePreview.ts): move to pure utils functions
